requirements.txt
# Core dependencies for NewAIBench framework
numpy>=1.21.0
pandas>=1.3.0
datasets>=2.8.0
pyarrow>=8.0.0

# Information Retrieval libraries
pyserini>=0.22.0
rank-bm25>=0.2.2
transformers>=4.21.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.3
hnswlib>=0.7.0

# Computer Vision for document images
timm>=0.9.0
torch>=1.12.0
torchvision>=0.13.0
pillow>=9.0.0

# OCR libraries
paddlepaddle>=2.4.0
paddleocr>=2.6.0
pytesseract>=0.3.10
easyocr>=1.7.0

# Machine Learning and experiment tracking
scikit-learn>=1.1.0
mlflow>=2.0.0
wandb>=0.13.0

# Data processing and storage
# sqlite3 - built-in module in Python
duckdb>=0.8.0

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.8.0
black>=22.0.0
isort>=5.10.0
mypy>=0.991
flake8>=5.0.0

# Documentation
sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0

# Utilities
tqdm>=4.64.0
click>=8.1.0
pyyaml>=6.0
python-dotenv>=0.19.0
